feat: improve swagger trait grouping
