+++ src/main/scala/com/tudux/taxi/http/Ping.scala
@@ -5,7 +5,7 @@ import akka.http.scaladsl.model.HttpResponse
 import akka.http.scaladsl.model.StatusCodes._
 import akka.http.scaladsl.server.Directives._
 
-case class Base(system: ActorSystem) {
+case class Ping(system: ActorSystem) {
   val routes = path("ping") {
     get {
       complete(HttpResponse(OK, entity = "pong"))
